Optimize distributions for when postfix! runtime checks are enabled
In most distribution code we're already doing a nil-check so we can
eliminate subsequent ones. For localAccess we know that the localArr
will not be nil because of implementation specifics. 14660 added some
compilation optimizations to eliminate runtime `!` checks, but those
don't help the distribution code much since we're mostly looking at
field accesses, which were not optimized.

With this I see no noticeable performance regressions for our single and
multi locale benchmarks and Arkouda if I manually enable postfix!
runtime checks.