Optimize large misaligned GETs in ugni
Optimize large misaligned GETs under ugni by using a single 0-copy RDMA
operation for the middle chunk of a misaligned GET. This improves 1MB
misaligned GET performance from 1.7 GB/s to 7.6 GB/s, which is just
slightly below the aligned 1MB speed of 7.9 GB/s

GETs on the Gemini NIC require 4-byte alignment for local address,
remote address, and the length. The Aries NIC dropped the 4-byte aligned
local address, but still requires aligned size and remote address. When
doing misaligned GETs, we have to peel off the misaligned remote source
address at the start, and the misaligned size at the end.

Previously we would transfer the middle section using bounce buffers
too, but this is unnecessary for Aries which doesn't require the local
target to be aligned. Update to just do a single RDMA GET of the middle
chunk for Aries. We can also do this for Gemini if the source/target had
the same misalignment and peeling off the source aligned the target.