Rename test `enum.chpl` to avoid naming conflict between module and `enum`
Like other tests renamed in #15150, this enum.chpl test needs to be
renamed to avoid its implicit module having a naming conflict with
the `enum` catch-all type.  This is similar to what happens if a test
is named 'int.chpl' today (and relies on implicit module names), and
while it'd be nice to improve the behavior, that wasn't the goal of
this effort.