Allow for regMemRealloc() support, and implement same in comm=ugni.
Adjust chpl_mem_array_realloc() to work if the comm layer supplies
registered memory reallocation support, and add the latter to comm=ugni.
Adjust the existing code in ChapelBase to adapt to the needs of this,
notably by passing both the old and new pointers and sizes to the
postRealloc() function.

While doing this, the need to add a realloc-specific postRealloc
function caused me to move the existing realloc() in chpl-mem-array.h so
as to achieve the order alloc(), postAlloc(), realloc(), postRealloc(),
free().