Change 'new owned C()' to 'new C()' in released examples
Now that 'new C()' is the same as 'new owned C()', let's use the
short form in the release/examples directory.  This occurred to
me after I saw people submitting examples using 'new owned ...'
in issues / PRs recently and wondered "Why are they doing that?"

As part of this effort, I did a little bit of wordsmithing and
formatting clean-up to the classes primer.