Remove .bad file for the new future I added
When testing on darwin, this bad file worked fine, but on linux I'm
getting additional output, and given the test's slightly weird
behavior, it seems likely/possible that the output may vary across
platforms.  So I'm going to remove the .bad given that
listOfTupleOfClass2.chpl has a .bad, and it seems likely that if/when
that case is fixed, this one will be examined and fixed as well.