Update submitted revcomps
revcomp.chpl and revcomp2.chpl both got retired because they allocated
a buffer the size of the input stream, which is no longer allowed by
the CLBG site.  Our new revcomp.chpl doesn't compile on master any
longer (though it seems like it ought to?), so I'm filing it here as a
future to keep this directory in sync with the shootout site.