Fix segfaults for reallocs of 0-element arrays
Engin pointed out that PR #15177 changed something about how we first
allocate arrays such that realloc was sometimes used on 0-length
arrays, and that this was causing segfaults in some tests.  Here, I'm
adding one additional test to the realloc path to make sure that the
array is non-empty.  This fixes at least some of the cases, I'm
going to run it through testing now.