Don't reallocate a DR array when its domain's value doesn't actually change
One gasnet test slipped past my notice today when making reallocation
of non-nilable classes illegal.  It turns out that the test was not
actually re-allocating an array, but simply assigning its domain the
same value again, which ought to be legal and should permit us to
avoid even going through the motions of doing the realloc.

Based on that insight, I'm adding a short-circuit check into the
dsiReallocate() code for DefaultRectangular arrays to avoid all the
work of pretending to reallocate in the event that we don't need
to.  This causes the test to pass and should also save some time
for people doing pointless domain reassignments.