Rewrite implementation to avoid reduce; add tests to see new code fires
By adding a reduce operation in my original implementation, I changed
things about module initialization order and dead modules that I
didn't intend.  I'd already been on the fence about whether the
reduction was a good idea or not (felt a bit like overkill), so here
I'm rewriting it to use a for loop to avoid those issues.

I also changed my existing tests to make sure normal realloc behavior
(errors and debug printing) wouldn't be detected in the event that the
module didn't change to lock this behavior in.