Disable ugni comm/compute overlap with --cache-remote
Comm/compute overlap with `--cache-remote` is currently leading to
segfaults. We don't fully understand the source of the problem, but for
now disable comm/compute overlap to workaround the problem. This will
hurt the performance of SSCA and microbenchmarks that oversubscribe
comm, but lower performance is better than segfaulting.

We'll look to fix this in the future, but for now this puts us in a
better state and I'd feel comfortable recommending `--cache-remote` to
users.