Parallelize PrivateDist scan
Optimize and parallelize scan for PrivateDist. The implementation just
gathers the PrivateArray into a local DefaultRectangular array, performs
the scan locally, and then scatters the results to the result array.
This is significantly faster than the default serial approach. It's 100x
faster at 512 nodes. PrivateDist scans are unlikely to be a performance
bottleneck for anybody but it was causing serialization warnings for
Arkouda.

Below are scalability numbers for PrivateDist scan. Results are a rate
(scans/sec) so higher is better. I'm just running:

```chpl
for 1..trials do
  var resEnds: [PrivateSpace] int = + scan numResults;
```

| nodes | before | now    |
| ----: | -----: | -----: |
|   1   | 56,400 | 44,000 |
|   2   |  8,200 |  7,500 |
|   4   |  3,700 |  5,000 |
|   8   |  2,000 |  3,800 |
|  16   |  1,200 |  2,800 |
|  32   |    640 |  2,200 |
|  64   |     90 |  1,750 |
| 128   |     40 |  1,400 |
| 256   |     15 |  1,000 |
| 512   |      6 |    600 |

There's a slight hit for single node since we're copying around a
little, but there's a 100x improvement at 512 nodes.