Remove studies/fft, which is entirely unused.
The tests under this directory seem to have been added (in 2008) with
the expectation that more work would be done.  The initial commit says
that the "README file is empty; its content will appear at a later time"
but that never did happen, nor did anything else.  Subsequent commits
did change the files here, but all of those just reflected the evolution
of the Chapel language or module code, not that of the tests themselves.
Checking the testing log files for this past Monday I see only "Working
on directory" and "No tests in directory" messages for this directory
and every subdirectory beneath it.  I think it's safe to delete it,
therefore.