Narrow all ddata's passed to init_elts()
Prior to this change, passing a ddata by `ref` to ddata_reallocate()
and then passing it along to init_elts() caused all calls to
init_elts() to use a wide rather than narrow ddata.  Worse, this even
affected programs that did not call ddata_reallocate() or generate it
in the generated code, I believe because of dynamic dispatch.  This
seems regrettable performance-wise, and was also subtle and hard to
track down.  It makes me wonder whether we should clone such functions
more aggressively to stop such cases from creeping in (and where else
that might result in performance savings).

Here, I work around that issue by having ddata_reallocate() behave
more like realloc() in that it takes in the old ddata and returns
the new one.

While debugging this, I also noted that we had an extraneous on-clause
in dsiReallocate() as the only place it's called is within
ChapelDistribution.chpl, which already contains the necessary
on-clause.  For that reason, I converted the on-clause into an assert
to reduce overheads.