When gathering PE args, give an "input file".
The behavior of `cc -craype-verbose -###` changed as of craype/2.6.5.3,
to wit: it seems to have begun requiring an input file in order to print
all the compilation and linking arguments it used to produce without
such a file.  Here, give an "input file" in the form of /dev/null.  (It
turns out that this also works before craype/2.6.5.3, so we just do it
in all cases.)