Couple more updates
Make validateRenamed a protected method on VisibilityStmt and shift
getSearchScope to being a pure virtual method on VisibilityStmt (because
validateRenamed relies on it, but that function has different behavior for use
versus import statement).

Also update ImportStmt::providesQualifiedAccess to only return true if qualified
and the renamed lists are both empty.