Deprecate `channel.error` and friends (#15297)
Deprecate `channel.error` and friends (#15297)

This PR deprecates methods that were used to check a channel for
errors in the world before #14739. These methods are no longer used
internally in IO code.

The Regexp module had one use of `channel.setError` that had to be
removed.

Adds one small test in `test/deprecated` to lock in the deprecation
warnings.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @mppf for review!