File futures showing inconsistency between transitive use of private members
This pair of tests shows that for a child that uses its parent which
uses its sibling, we are inconsistent in that we do resolve functions
but don't resolve variables.  Rather than calling either of these
behaviors correct, I'm filing both as futures until we can sort out
what ought to happen.

One argument would be "default `use` is not transitive, so it makes
sense that a module using `parent` would not see symbols that it gets
access to via its `use`s.  Moreover, such a module can always `import`
or `use` the module or symbols that it wants directly.

Another argument would be that child modules have a special
relationship with their parent, so should be able to access things
that it can.

A counterargument might be "If a parent wanted to use one of its child
module's symbols without making all its children aware of them, how
could it do that if its uses were transitive to children in this way?"