Split copyright lines to preserve historical Cray copyright
While talking with the HPE open-source board for other reasons,
I realized that rather than having HPE just absorb the Cray
historical copyrights, it should really be appended to them.
This PR splits all the copyright lines into two, one showing
the historical Cray copyright and a second asserting that HPE
holds the copyright as of 2020.