Only compute LICM aliases once
LICM computes all aliases in a given function. Previously, it was
computing those aliases for each loop in that function, but that
wastes a lot of time, especially for large functions with many loops.
Only compute the aliases once per function and pass them to the loops.
This halves LICM for Arkouda.