Add additional constraint to shift expressions in spec (#15304)
Add additional constraint to shift expressions in spec (#15304)

This PR updates the bitshift section of the spec to reflect the
changes made in #14391. It adds a constraint that "restricts the
value of the second operand to be less than the number of bits
required by the type of the first".

---

Testing:

- [x] Make sure that docs render locally

---

Thanks to @e-kayrakli for review!