Correct `list.last` bug due to ref and on clause, add tests (#15369)
Correct `list.last` bug due to ref and on clause, add tests (#15369)

This PR fixes a rather nasty bug in `list.last` caused by a
combination of `on` clauses and my own misunderstanding of Chapel
references (they can't be reassigned!). It also adds tests to lock
in the correct behavior for `list.last` and `list.first` (which
also exhibited the bug).

Resolves #15366.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] library/standard/List/listRefBugMultiLoc.chpl on darwin when
      CHPL_COMM=gasnet

---

Thanks to @ben-albrecht for review!