LinkedList: Remove superfluous ref this intent from append() and remove()
These methods modify this, so have ref this intent by default.

None of the other methods in the module have an explicit ref intent,
including the ones that modify this, such as the other versions of
append().  So remove these two for consistency, and so readers don't
think they're necessary.