Oops, fix a subtle error
Previously, we were declaring domains using ranges like `0.. #dims(0)`
which creates an int(64) range, then applies a 32-bit count to it,
resulting in an int(64) range, which can be assigned to an int(64)
domain.  When I switched to `0..<dims(0)` this changed it to an
int(32) range, which was not legal to assign to an int(64) domain (at
present).  However, since this test is skipped by default, I didn't
catch it at first...