Remove flow control associated with RMA/AMO EP completion handling.
There were several instances of divergent flow control based on whether
the target-side RMA/AMO EP used a CQ or a counter for completions.
Here, get rid of nearly all of that by recording the fid and a pointer
to an appropriate completion-consuming function at the time we open the
CQ or counter, and using those when we add it to and remove it from the
poll set, close it, or consume completions through it.