Remove flow control associated with transmit EP completion handling.
There were several instances of divergent flow control based on whether
the source-side transmit contexts (endpoints) used CQs or counters for
completions.  Here, get rid of nearly all of that by recording the fid
and a pointer to an appropriate completion-consuming function at the
time we open each transmit CQ or counter, and using those when we add it
to and remove it from the poll set, close it, or consume completions
through it.

While here, also add a pointer to a progress function for each transmit
context, and refer these to the appropriate functions for regular worker
tx contexts and AM handler tx contexts.  For the former we only need to
be responsible for progress on the tx context itself, while for the
latter we also need to be responsible for progress on the inbound AM
message endpoint and the RMA/AMO endpoint, at least if we're not set up
for automatic progress.