Fix off-by-one errors in UnitTest
Somehow these slipped by me, but one of my previous commits broke
the way < and > unit tests print out their output without my
noticing.  This is because the test to see whether we were at the
end of the array or not was comparing whether the index (now 0-based
was equal to the size).