Switch fannkuch to a more explicit array type form
This changes fannkuch in a different way than my previous attempt.
Rather than updating the iterator to yield 0-based factorial values,
we change the array to be 1-based and more explicit.  This is a
great motivator for the `var fact: [1..] = ...` form in which
both size and element type are inferred rather than required to be
specified explicitly.