Fix find()-related bug in sort logic
The sort performance test had a subtle bug in which it relied on
'if find(...)' idioms to determine which sorts to run.  With strings
and bytes going to 0-based indexing, and find returning -1 on
no-match finds, this caused all sorts to be run rather than just one
in performance runs, causing timeouts and bad data.  Here, I'm
switching the find calls to compare to -1 and also updating the
correctness tests to make sure the output works correctly by only
squashing timing output, not all output.