Switch from supporting cce-classic to cce-clang
Add support for the newer clang-based versions of the cray compiler,
removing support for the classic cray compiler. It's feasible that we
could support both, but we have problems with 9.x.x-classic and older
versions are quickly getting phased out, so unless a user requests it,
let's just support the clang-based cce.

To support it we just include the clang makefile and remove an
cce-classic specific flags we were throwing before.