Add future for het. loop over varargs tuple with string and --baseline
This adds a future showing that a loop over a heterogeneous varargs
tuple containing a string fails when compiled with --baseline, where
our normal tests of loops over heterogeneous tuples don't.  This
showed up when the procedures.chpl primer was rewritten to do just
that in --baseline testing, so suppress that test for --baseline
until the future is resolved.