Remove test/functions/iterators/vass/TIMEOUT
Remove an old TIMEOUT file. It was originally added in b05104a8d3 to get
a compiler hang to fail faster, but it was applied to the entire
directory instead of just the failing test. This has caused timeouts
under valgrind for other tests in this directory. Remove the timeout and
add a .bad to the failing test, since it is no longer hanging.