Make random streams use 0-based indexing
This PR makes RandomStreams count from 0 rather than 1 in terms of
routines like getNth() or skipToNth().  This is largely not a
backwards-breaking change unless you happen to care what numbers
are returned for a given random stream index since we're only
expanding the set of indices for which a random number can be queried.

The approach I took here for the sake of expediency was to have the
random stream continue to store its state as before, and to just
adjust the values by 1 at the interface boundaries that refer to
a specific position.

The main work was updating the documentation and adjusting our tests
which are reasonably sensitive to which values they get from the
random streams