Reshuffle information on import and use statements in the spec
Moves the majority of the information from the Statements chapter of the spec
into the Modules chapter of the spec.  Removes some trailing whitespace, and
updates the link from the Enums section to point to both spots instead of just
the shorter section (so that readers can get to all the information more
quickly).