Fix race-y out-of-bounds tuple indexing error in Sort.chpl
This fixes an out-of-bounds error in tuple indexing that only happens
occasionally in practice (~1% of the time when compiling with --verify
on my Mac) and as a result slipped through testing until now.  It's
also subtle in that it's done indirectly by a routine that takes the
thing to be indexed into (often an array, but this time a tuple) and
its bounds as arguments, so there was indirection involved.  Searching
through the code, I'm not seeing other obvious places that would be
running afoul of the same issue.