Standardize naming, for runtime-maintained per-task information.
The names of struct members and functions having to do with per-task
information mainted by the runtime has been somewhat irregular and not
entirely descriptive in some cases.  Here, impose more order.

These have to do with space maintained by the runtime but with contents
managed by the Chapel module code, containing per-task information such
as the task's serial state:
  chpl_task_ChapelData_t --> chpl_task_infoChapel_t
  chpl_task_getChapelData() --> chpl_task_getInfoChapel()
  chpl_task_getBundleChapelData() --> chpl_task_getInfoChapelInBundle()
  state --> infoChapel (member in various task/comm arg bundle structs)

These have to do with per-task information managed by the runtime and
not of concern to any other code, such as the per-task comm cache:
  chpl_task_prvData_t --> chpl_task_infoRuntime_t
  chpl_task_getPrvData() --> chpl_task_getInfoRuntime()

While here, also remove chpl_task_getPrvBundle().  This was a building
block which allowed implementing chpl_task_getInfoChapel() in common
code, so that tasking layers didn't have to.  But that wasn't a win,
because the tasking layers had to implement chpl_task_getPrvBundle()
themselves anyway since it was inherently implementation-specific.  So
here, implement chpl_task_getInfoChapel() in the tasking layers and get
rid of the unnecessary building-block function.