Make --library-python defaults more Python friendly (#15475)
Make --library-python defaults more Python friendly (#15475)

This PR adjusts the default output directory for `--library-python`
compiles to be the same as the Python module name. It adjusts the
compopts for the tests in `interop/python`, forcing the output
directory for old tests to be `--library-dir=lib` to avoid changing
old tests.

This PR also modifies `--library-python` compiles to add a simple
`__init__.py` file in the output directory. This enables a user to
import the output directory as a Python package.

The initializer file also registers `chpl_cleanup()` to be called at
program exit. If imported as a package, the user only has to call the
`chpl_setup()` routine themselves.

The contents of the initializer file are fixed and do not vary from
single to multilocale compiles.

The library technote has been updated to reflect these new features.

Resolves #13880.
Resolves #13882.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] interop on linux64 when CHPL_COMM=none, cython, numpy
- [x] interop on linux64 when CHPL_COMM=gasnet, cython, numpy
- [x] interop on darwin when CHPL_COMM=none, cython, numpy
- [x] Generate doc changes

---

Thanks to @lydia-duncan and @ben-albrecht for review!