Automatically remove clock skew warnings in sub_test
We occasionally see clock skew warnings in our testing. This warning is
from `make`ing our generated code. The warning is harmless, but it
creates testing noise. We used preexec-for-gmake-clock-skew-warning in
some configurations to quiet this, but we didn't use it everywhere and
developers running testing could still see the warning. Update sub_test
to automatically remove the warning instead of relying on a preexec.

The clock skew warning is innocuous, so there's no harm in hiding it