Fix futures for tuples of classes, add future for param folding of if expression (#15541)
Fix futures for tuples of classes, add future for param folding of if
expression (#15541)

Several futures filed for #15394 failed due to incorrect folding of a
param if expression which resulted in a compiler error. This PR
adjusts the tests filed for tuples in #15394 to sidestep the param
if expression entirely. It does so by making use of a function
overload with a where clause. All tests in `test/types/tuple/types`
now pass.

---

Thanks @ben-albrecht for signing off!