Allow the gdb and lldb lview command for Chapel AST nodes to accept consts
Prior to this change, lview/list_view would accept non-const versions and print
the contents of the AST node nicely, but would throw up its hands if it was
passed a const version of the same node.  This was noticed when trying to print
the argument for ImportStmt::providesNewSymbols.

By its definition, lview/list_view should never modify the contents of the nodes
that were passed to it, so it should be safe to change list_view and its helpers
to accept const arguments (where non-const value would just be const for the
duration of the function).  Based on checking const and non-const AST nodes
with this change, this appears to be accurate.

Details:
--------
In order for this update to work, I had to make a couple of related AST changes.

- a copy of isForallLoopBody was made specifically for const arguments, named
isConstForallLoopBody.  It is otherwise identical to isForallLoopBody - I was
concerned that returning a const ForallStmt would have wider consequences in the
compiler, so sidestepped the issue.

- There are some slight whitespace changes

- I marked ArgSymbol's intentDescString method as const, since it does not
change the contents of the node, and did the same for SymExpr's symbol method.

- list_sym now takes a const symbol and uses toConst* instead of to*.  I also
changed some uses of to* where the result was dropped on the floor to instead
rely on is* (e.g. a toFnSymbol call which did not save the result was turned
into isFnSymbol).

- Similarly, block_explanation, forall_explanation_start, forallPreamble, and
list_line now take a const and use toConst* instead of to*

- forallPostamble required const versions of ForallStmt's inductionVariables and
iteratedExpressions methods.  As with isForallLoopBody, I was similarly
concerned about the wider consequences of returning a const AList& to other
places in the compiler.

- usePostamble and importPostamble needed to use const_iterators instead of just
iterators in addition to const arguments.

- list_ast was updated similarly to list_sym - const arguments, uses of toConst*
instead of to*, and uses of is* when the result of to* would be dropped on the
floor.  This function also needed an update to conditionally set a variable's
value that will now be const.

- list_view(int id) now saves the result of aidWithError as a const, and
list_view(BaseAST* ast) and list_view_noline now take a const and use is*
instead of dropping the result of to*.

- AST_CHILDREN_CALL needed to be updated due to code in the BlockStmt branch
that forced it to not accept const BlockStmt*.  Thankfully, the work was mostly
done to allow this conversion to occur - headers for functions that wrap the
needed BlockStmt method calls were already present in the file, it was just a
matter of using them instead of the methods and removing the local variable that
was causing problems.

Testing:
- verified the behavior of const UseStmts, non-const UnresolvedSymExprs,
non-const BlockStmts and non-const type definitions.

Going to run a full paratest with futures to be certain there weren't unintended
consequences.