Make a "friendly" error for sparse arrays of non-nilable classes
Sparse arrays are (typically) so dynamic by nature that elements are
added (as well as removed), which poses a problem for our current
implementations when the element type is non-nilable since the new
array values need a default value up until the time they are assigned.
Previously, the internal implementations of sparse arrays would generate
an error; here, I'm adding an explicit check and error for this case.

In reality, this check should not be looking merely for a non-nilable
class eltType, but any eltType that doesn't have a default value; but
due to issue #14854, this isn't possible at present, so I'm taking the
cheap / lazy approach instead.  Once #14854 is resolved, this check
for isNonNilableClass() should be updated to !isDefaultInitializable().