Update three places to use ImportStmt's providesUnqualifiedAccess method
This will make the code more robust to potential changes in the future
(especially w.r.t. if we enable importing both the module and symbols in it in
the same {} list).

It also fixed a bug where we were giving a less useful error message in the case
where the last symbol before an unqualified list was not a module, but the
unqualified list only contained renames.