Update validateNamed and validateUnqualified when complaining about private
Private symbols shouldn't be allowed in the only/except lists of use statements
or in the unqualified lists of imports, unless we are in a child scope of where
the private symbol was defined.  To that end, check the symbol's visibility
before erroring.

Note: this may not solve the problem for renamed symbols, I haven't written a
test for that case just yet.  I also need a test case to ensure we can't import
a private module living in a public module we've already used.