Enable import statements to use super to access a symbol in their parent module
Prior to this change, `import super.foo;` would only work if foo was a module.
However, import statements are allowed to indicate a single symbol within a
module for unqualified access, and since we require an import or use of parent
modules to access their symbols now, it makes sense to enable a faster way of
indicating a symbol in the parent module (by using `super` instead of the full
module path).