Use workaround to enable lifetime checking for lists of borrowed classes (#15575)
Use workaround to enable lifetime checking for lists of borrowed
classes (#15575)

This PR adds a small workaround to the List module that "encourages"
the compiler to perform lifetime checking for methods such as
`list.append()`.

This workaround is required because lists use `_ddata` for their
backing store instead of an array. When it comes time for the compiler
to evaluate lifetime clauses, a lifetime clause is only considered for
evaluation if both formals in the clause are or refer to a borrowed
class. The `_ddata` type is marked as unsafe by the compiler and is
skipped. Because the compiler is unable to infer that a
`list(borrowed C)` contains a borrowed class, lifetime clauses on list
methods such as `this < x` are not evaluated.

To resolve this issue, I added a dummy field to the `list` type. If
the `eltType` of a list is a borrowed class, this field will have the
type `eltType?` (a nilable `eltType`) and is default initialized to
`nil`. Otherwise this field will have the type `nothing` and is folded
away.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Reviewed by @mppf.