Stop updating the mod to search based on the re-exported symbol
It was only catching the last symbol that matched, making the value returned not
useful in the case where multiple imported scopes provided candidates - we were
limiting the scope we'd look at to this match, but this meant that we couldn't
find versions in other scopes.  Instead, just pass in the named module, we
should update function resolution to instead search from there.