Add a map for re-exported symbols to visible function determination
Only used when going directly into a scope because of a module prefix.  Instead
of only allowing symbols directly defined in the module, we will now check if
this scope has already been checked for re-exports with that name.  If not, we
will follow all public import statements in that scope, as such statements
re-export symbols.  We'll update our visible function block to store any results
we find for next time.

This seems to resolve the future I was looking at.  I need more extensive
testing to be certain, but first I'd like to run a paratest and see what it
tells me