Adjust insertCasts to handle PRIM_ASSIGN in addition to PRIM_MOVE
Code for doing "move initialization" in internal modules
is currently written with PRIM_ASSIGN since that is more or
less what is needed. However the test

  test/arrays/ferguson/array-of-owned-no-error

revealed that when the types were different we were not performing
a cast as we should. An alternative strategy for solving the problem
with that particular test would be to adjust convertToCanonical in
DecoratedClassType.cpp to convert owned/shared nilable -> owned/shared
non-nilable.

Additionally, getting this test working required changing
the return-by-ref to avoid creating a ref temporary in this case.