Set: op= in-place, write some op in terms of op= for terseness
Write the |=, +=, -=, and ^= to operate on lhs in-place rather than
using a temporary and replacing all of lhs with the temporary.  (The
temporary they were using was inside the |, +, -, or ^ operator they
invoked.)

This allows |= / += to not invalidate references to existing elements.
The other operations still do since they necessarily remove elements
from lhs.

Rewrite &= to be in-place when it's efficient to do so.  It still
invalidates references due to .remove().

Write | and ^ in terms of |= and ^= to reduce code duplication.

& and - weren't as amenable to writing in terms of &= and -=, for
efficiency reasons.

Since & is symmetric, and only has to iterate over one of its
arguments, choose to iterate over the smaller argument.

Add a test of each op= to the test of its op.

Correct name of module in module-ending comment.